fix(#36): stops deletion of hot-keys for CMDs for other plugins
This fixes an issue in `cleanHotKeys()` that had bad cmd-key deletion
logic; because this was called by `onLoad()` it would delete cmd hotkeys
managed by plugins that were loaded _before_ toggl-list.

It is pretty clear that the intended logic in `clearHotKeys()` was to
delete toggle-list's hot-keys, not all other keys _except_
toggle-hotkeys.

This fixes the incorrect logic when using 'Array.find'; it was `!find`
when `find` was intended. This inverted logic caused all
*non*-toggle-list keys to be deleted.